requirements.txt
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.30.0
sqlmodel>=0.0.22
psycopg2-binary
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
httpx>=0.27.0
